requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
boto3==1.35.64
botocore==1.35.64
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.1.0
django-concurrency==2.6.0
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
pydub==0.25.1
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyVirtualDisplay==3.0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rpds-py==0.21.0
s3transfer==0.10.3
selenium==4.27.1
setuptools==78.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==14.1
whitenoise==6.8.2
wsproto==1.2.0
zoom-meeting-sdk==0.0.23
cachetools==5.5.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-texttospeech==2.24.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
rsa==4.9
ruff==0.9.6
durationpy==0.9
kubernetes==32.0.0
stripe==11.6.0
tldextract==5.3.0